refactor: entity 클래스 패키지 구조 변경
