+++ src/main/java/jpa/favorite/FavoriteRepository.java
@@ -1,8 +1,6 @@
-package jpa.dao;
+package jpa.favorite;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import jpa.domain.Favorite;
 
 public interface FavoriteRepository extends JpaRepository<Favorite, Long> {
