+++ src/main/java/jpa/line/LineRepository.java
@@ -1,8 +1,6 @@
-package jpa.dao;
+package jpa.line;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import jpa.domain.Line;
 
 public interface LineRepository extends JpaRepository<Line, Long> {
 	Line findByName(String name);
