+++ src/main/java/jpa/station/StationRepository.java
@@ -1,10 +1,8 @@
-package jpa.dao;
+package jpa.station;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import jpa.domain.Station;
 
 public interface StationRepository extends JpaRepository<Station, Long> {
